kind of working

+++ history.py
@@ -9,7 +9,7 @@ class History:
         self.screen_width = screen_width
         self.screen_height = screen_height
 
-        self.history = np.zeros((self.history_len, self.screen_height, self.screen_width), dtype=np.float16)
+        self.history = np.zeros((self.history_len, self.screen_height, self.screen_width), dtype=np.uint8)
 
     def add(self, screen):
         self.history[:-1] = self.history[1:]
